combine script files. Add additional unit tests
