+++ Shadowrun5thEdition/Shadowrun5thEdition.pug
@@ -651,11 +651,11 @@ script(type="text/worker")
 	include src/js/Classes/PrimaryWeapon.js
 	include src/js/Classes/PrimaryArmor.js
 
+	//- Functions that do most the work.
+	//- These two functions files have unit tests for Mocha
 	include src/js/processingFunctions.js
+	include src/js/shadowrunFunctions.js
 
-	include src/js/newScripts.js
-	
-	//Old functions working to become obsolete
 	include src/js/scripts.js
 
 	include src/js/versioning.js
@@ -663,7 +663,6 @@ script(type="text/worker")
 	include src/js/eventhandlers.js
 
 	include src/js/importer.js
-
 	include src/js/importerUI.js
 
 //- CHUMMER IMPORTER
